+++ lib/sys_types.dart
@@ -1,0 +1,6 @@
+/// POSIX System Data Types
+typedef pid_t = int;
+typedef mode_t = int;
+typedef off_t = int;
+typedef ssize_t = int;
+typedef useconds_t = int;